refactor: 責務分離のためファイル分割
